Decompose the network config from the ServerOptions.
This will be re-used as I split the proxy out of the main LdapServer. This consolidates the shared usages into a trait and a DTO. Shared pieces between the two will use an interface instead of the main options object. Or they will receive the network config DTO.
